Add tap-hold behaviour to TAB button
TAB on tap, ALT on hold. Makes easier to use CMD + ALT + LEFT/RIGHT
shortcut, which is used in Chrome for tab switching, and in JetBrains
products for Back/Forward.

Another reason for this is that it is quite uncomfortable to reach for
ALT with a thumb. Such ALT placement might help with that.