Move symbols from Lower layer to Raise
The goal is to make better use of layers. Both Lower and Raise have an
F-row, what a waste of space. Symbols are split between two layers,
yuck. There are also few duplicate keys that I already added to other
layers. This commit fixes all of that, leaving Lower with only F keys.
There might be a better place for them too though.